+++ src/Utility/BackupManager.php
@@ -91,8 +91,6 @@ class BackupManager
 
         return (new Collection($Finder))
             ->map(fn (SplFileInfo $File): array => [
-                /** @todo remove `filename` in version 2.14.0 */
-                'filename' => $File->getFilename(),
                 'basename' => $File->getBasename(),
                 'path' => $File->getPathname(),
                 'compression' => Compression::fromFilename($File->getFilename()),